+++ src/main/java/calculator/OperatorMethod.java
@@ -1,4 +1,4 @@
-package calculator.domain;
+package calculator;
 
 public interface OperatorMethod {
     int operate(int num1, int num2);
